Now it works in a true multithreaded fashion! yaay! Also, instead of rolling that weird implementation of thread management, now std::async and std::future is used

+++ src/ExecuteCommand.cpp
@@ -1,15 +1,17 @@
 #include <cstdio>
 #include <vector>
 #include <string>
+#include <atomic>
 
 #include "../include/Core.hpp"
 #include "../include/Environment.hpp"
 #include "../include/FSFuncs.hpp"
+#include "../include/StringFuncs.hpp"
 #include "../include/DisplayFuncs.hpp"
 
 #include "../include/ExecuteCommand.hpp"
 
-std::vector< Exec::Internal::Result > Exec::Internal::threadinfo;
+std::atomic< int > Exec::Internal::threadctr( 0 );
 
 int Exec::ExecuteCommand( const std::string & cmd, std::string * err )
 {
@@ -31,28 +33,30 @@ int Exec::ExecuteCommand( const std::string & cmd, std::string * err )
 	return Core::ReturnVar( res );
 }
 
-void Exec::MultithreadedExecute( const std::string & cmd, int thread, const std::string & src )
+Exec::Internal::Result Exec::MultithreadedExecute( const std::string & cmd, const std::string & src )
 {
-	std::string tmp_file = Core::TMP_FILE + std::to_string( thread );
+	std::string tmp_file = Core::TMP_FILE + std::to_string( Internal::threadctr );
 	std::string cmdfinal = cmd + " 2>" + tmp_file;
 
 	// No log section here because multiple threads would spawn multiple sections
+	Core::logger.AddLogString( LogLevels::ALL, "Exec::MultithreadedExecute: Executing: " + cmdfinal );
 
-	Core::logger.AddLogString( LogLevels::ALL, "Exec::MultithreadedExecute: Executing: " + cmd );
-
-	int res = std::system( cmdfinal.c_str() );
-
+	int res;
 	std::string err;
+
+	FILE * pipe = popen( cmdfinal.c_str(), "r" );
+
+	if( !pipe ) {
+		return { 1, "Pipe open failed", src };
+	}
+
+	res = pclose( pipe );
 
 	if( FS::LocExists( tmp_file ) ) {
 		err = FS::ReadFile( tmp_file );
 		FS::DeleteFile( tmp_file );
 	}
 
-	// res must be last because i am not using mutex, therefore
-	// once res is set to some value, the main thread will do it's job
-	// thereby corrupting the entire threadinfo
-	Internal::threadinfo[ thread ].err = err;
-	Internal::threadinfo[ thread ].src = src;
-	Internal::threadinfo[ thread ].res = res;
+	--Internal::threadctr;
+	return { res, err, src };
 }
